move mathitems redux state to directory
