+++ client/src/features/sceneControls/mathItems/index.tsx
@@ -1,20 +1,9 @@
 import MathItem from "./MathItem";
-import type { MathItemsState } from "./mathItems.slice";
-import mathItemsSlice, {
-  useMathItem,
-  selectMathItems,
-  selectSubtree,
-} from "./mathItems.slice";
+import type { MathItemsState } from "./mathItemsSlice";
+import mathItemsSlice, { reducer, actions, select } from "./mathItemsSlice";
 import { MathContext } from "./mathScope";
 
 export { default as FolderWithContents } from "./FolderWithContents/FolderWithContents";
 
 export type { MathItemsState };
-export {
-  MathContext,
-  MathItem,
-  mathItemsSlice,
-  useMathItem,
-  selectMathItems,
-  selectSubtree,
-};
+export { MathContext, MathItem, mathItemsSlice, reducer, actions, select };
