+++ client/src/features/sceneControls/mathItems/mathItemsSlice/index.ts
@@ -1,0 +1,8 @@
+import * as select from "./selectors";
+import type { MathItemsState } from "./mathItems.slice";
+import mathItemsSlice, { actions, reducer } from "./mathItems.slice";
+
+export { select, actions, reducer };
+
+export type { MathItemsState };
+export default mathItemsSlice;
