+++ client/src/features/sceneControls/mathItems/templates/ItemTemplate.tsx
@@ -11,7 +11,7 @@ import { useAppDispatch, useAppSelector } from "store/hooks";
 
 import ColorStatus from "../ColorStatus";
 import FieldWidget, { useOnWidgetChange } from "../FieldWidget";
-import { actions, selectIsActive } from "../mathItems.slice";
+import { actions, select } from "../mathItemsSlice";
 import { usePopulateMathScope } from "../mathScope";
 import { testId } from "../util";
 import CloseButton from "./CloseButton";
@@ -36,7 +36,7 @@ const ItemTemplate = <T extends MIT>({
   const onWidgetChange = useOnWidgetChange(item);
   usePopulateMathScope(item, config);
   const dispatch = useAppDispatch();
-  const isActive = useAppSelector(selectIsActive(item.id));
+  const isActive = useAppSelector(select.isActive(item.id));
   const remove = useCallback(() => {
     dispatch(actions.remove({ id: item.id }));
   }, [dispatch, item.id]);
